fix(#1122): cast d (usize) to f64 in matern triplet t_psi/t_psi_psi (E0277)
The fix-1122 triplet-identity derivative lines `t_psi = (lap_psi - d*ratio_psi)/r2`
multiplied `d` (centers.ncols(), usize) by f64 ratios. Add `(d as f64)` at both
sites (part_003.rs:1145-1146). Only `d * ratio` sites in the block; the other
`d`-arithmetic is `p*d*d` array sizing (usize), unaffected.

+++ src/terms/basis/split_parts/part_003.rs
@@ -1142,8 +1142,8 @@ fn build_matern_operator_penalty_psi_derivatives(
             } else {
                 let r2 = r * r;
                 (
-                    (lap_psi - d * ratio_psi) / r2,
-                    (lap_psi_psi - d * ratio_psi_psi) / r2,
+                    (lap_psi - (d as f64) * ratio_psi) / r2,
+                    (lap_psi_psi - (d as f64) * ratio_psi_psi) / r2,
                 )
             };
             d0_raw[[k, j]] = phi;
